+++ Sources/SwiftExtensions/NumberFormatter+Extension.swift
@@ -16,4 +16,12 @@ extension NumberFormatter {
         percentFormatter.multiplier = 100
         return percentFormatter
     }()
+
+    // Example for English locale: "1,234.5" -> 1234.5
+    static let localizedNumberFormatter: NumberFormatter = {
+        let formatter = NumberFormatter()
+        formatter.numberStyle = .decimal
+        formatter.locale = Locale(identifier: OAUtilities.currentLang() ?? Locale.current.identifier)
+        return formatter
+    }()
 }
